requirements.txt
mysql-connector-python==8.0.30
requests==2.28.1
dateparser==1.1.1
beautifulsoup4==4.11.1
